refactored out signature generation from header generation
